Add ip address cache
Fix #1. This commit introduces an optional ip address cache which stores
the optained IPv4 and IPv6 addresses between two runs of the program.
This reduces netword usage as the DNS records from netcup only need to be
fetched when the current ip addresses differ between those from the last
run.